Add time span validation

+++ backend/RSService/BusinessLogic/IRSManager.cs
@@ -9,7 +9,7 @@ namespace RSService.BusinessLogic
     public interface IRSManager
     {
         IEnumerable<Event> CreateAvailabilityEvents(DateTime startDate, DateTime endDate, int[] hostId, int[] roomId);
-        int GetTimeSpan(DateTime start, DateTime end);
+        double GetTimeSpan(DateTime start, DateTime end);
         int GetAvailableTime(int attendeeId, DateTime startDate);
         bool CanCancel(DateTime startDate, DateTime endDate, int roomId, int attendee);
     }
